ASSET-26 Move terminal types into the models/terminals
