+++ libraries/models/currency.py
@@ -1,7 +1,7 @@
 from pydantic import NonNegativeInt
 
-from libraries.models.address import Address
-from libraries.models.id import Id
+from libraries.models.terminals.address import Address
+from libraries.models.terminals.id import Id
 from libraries.models.templates.camelcase_model import CamelCaseModel
