+++ libraries/models/enum_info.py
@@ -1,5 +1,5 @@
-from libraries.models.description import Description
-from libraries.models.id import Id
+from libraries.models.terminals.description import Description
+from libraries.models.terminals.id import Id
 from libraries.models.templates.camelcase_model import CamelCaseModel
