+++ libraries/models/terminals/description.py
