+++ libraries/models/terminals/image_type.py
