+++ libraries/models/terminals/network_type.py
