+++ libraries/puller/getters/explorer_getter.py
@@ -7,8 +7,8 @@ from prompt_toolkit.completion import WordCompleter
 from prompt_toolkit.document import Document
 from prompt_toolkit.validation import Validator, ValidationError
 
-from libraries.models.id import Id
 from libraries.models.reference import Reference
+from libraries.models.terminals.id import Id
 
 
 class ExplorerValidator(Validator):
