+++ libraries/puller/getters/id_getter.py
@@ -8,7 +8,7 @@ from prompt_toolkit.document import Document
 from prompt_toolkit.validation import Validator, ValidationError
 from pydantic import ValidationError as PydanticValidationError
 
-from libraries.models.id import Id
+from libraries.models.terminals.id import Id
 
 
 class IdValidator(Validator):
