+++ libraries/puller/token_pullers/token_puller_dexguru.py
@@ -4,7 +4,7 @@ from bs4 import BeautifulSoup
 from requests import post, get
 from yarl import URL
 
-from libraries.models.address import Address
+from libraries.models.terminals.address import Address
 from libraries.models.network import Network
 from libraries.puller.getters.token_count_getter import TOKEN_COUNT_PER_PAGE
 from libraries.puller.token_pullers.token_puller_abstracted import TokenPullerAbstracted
